Add Discordrb::Middleware
This adds `Discordrb::Middleware` as an alternative to `EventContainer`
that allows users to build entirely custom chains of reusable functions
on top of event handlers.

This allows for things like custom event filters, or something like a
message parser that forwards parsed artifacts to the event handler
through a state variable.

Middleware enabled event handlers can be used on instances of `Bot` by
executing:

  require "discordrb/middleware"
  Discordrb::Bot.include(Discordrb::Middleware)

Or with the container pattern:

  module MyHandlers
    extend Discordrb::Middleware

    message(middleware_a, middleware_b, from: "z64") do |event, state|
      # ..
    end
  end

  bot.include!(MyHandlers)

Included is an API for registering keyword arguments of event handlers
to return instances of middleware for easy re-use across an application.
(`Discordrb::Middleware::Stock.register`)

Provided is `MessageFilter`, which is bound to the `bot.message`
handler, which creates the standard set of attribute matching features
that base `Events` system provides.

See documentation and examples for further details.