Exclude lib/discordrb/middleware from Documentation cop
As discussed in Discord, we determined the changes to silence the
Documentation cop for this directory would be overall harmful to the
code style of middleware/ files. We already get docs coverage CI from
YARD, but it isn't desireable to completely disable rubocop's
Documentation across the entire project just for this.